ACPI: battery: Report POWER_SUPPLY_STATUS_NOT_CHARGING
When there is AC power and the current going out of the battery is 0,
report the status as not charging.

Please note we do this only after checking all the other possible states
and when the remaining option is to return an unknown status. In that
case, and only in that case, we check for
POWER_SUPPLY_STATUS_NOT_CHARGING using the AC power info and rate_now.

Also, note that the code in acpi_battery_is_not_charging() is similar to
acpi_battery_handle_discharging(), but code is not reused as they are
semantically quite different: this is not a quirk nor some bug while
discharging, and therefore the return values both function have are
different. We can try to force them to be similar, but that will mess
the code in acpi_battery_get_property() and seems too stretch.

Signed-off-by: Rodrigo Campos <rodrigo@sdfg.com.ar>